Fix for schema.ts, Drizzle, create user: set default id
``The issue occurs when logging with a new user with oAuth by discord, no user_id is created by drizzle ORM. where as Prisma supports CUID.  as Drizzle requires not-null value without creating one on account creation. current solution i can think of is to assign a default value to user id.```. Here is an update I made to "schema.ts" for SQLite, which it is currently using: ```//for SQLite:
export const users = createTable("user", {
  id: varchar("id", { length: 255 })
    .notNull()
    .primaryKey()
    .$defaultFn(() => crypto.randomUUID())
}
```